fix definition of props and state for @DiffSectionSpec
Summary:
D5493757 did a poor job creating an accurate DiffSectionSpecModel. DiffSectionSpec is just like everyone else _except_ the Props and state in the OnDiff delegate method must have a Diff wrapper type.

This diff changes the DiffSectionSpecModel to reflect the proper differences between itself and the rest of the litho plugins.

Reviewed By: IanChilds

Differential Revision: D5591532

fbshipit-source-id: 1286cce622b0490db6188c99f8a8d8e5c22f4b76