Reset DiffNode#mBorder in release()
Summary: We were getting a very weird bug when using borders where we'd try to do a shouldUpdate with a non-border Component as the previous component. The problem is we were recycling the DiffNode but not resetting the border property, which then pointed to a re-used LayoutOutput containing some arbitrary component.

Reviewed By: IanChilds

Differential Revision: D5639749

fbshipit-source-id: b4fae18b89caa41908d1a814779492445ca76ac8