Add Condition wrapper for HamcrestMatchers
Summary:
AssertJ Conditions and Hamcrest Matchers are roughly (pardon my mathematical
imprecision) isomorphic, but sadly Java doesn't let us express that. So instead,
we wrap the Matcher in a Condition and transplant the error message.

That way we can use a bunch of the existing Matchers while writing our tests.
Example, as always, provided in the sample code.

The result is - if I may say so - quite cool:

```
assertThat(c, mComponent)
    .has(subComponentWith(c, text(containsString("10"))));
```

Reviewed By: IanChilds

Differential Revision: D5630679

fbshipit-source-id: a5f4d917d408eda00344411a0de20f65b003a49d