Add support for ViewRenderInfo in RecyclerBinder
Summary:
Adding support for rendering views directly in `RecyclerBinder`. This diff will allow app developers to specify how views will be created and bound, and the framework takes care of the rest. The API looks like the following:
```

  private static ViewCreator VIEW_CREATOR = new ViewCreator() {
    Override
    public  View createView(Context c) {
       return new MyCustomView(c);
    }
  };

  OnEvent(RenderEvent.class)
  RenderInfo onRenderEdge(
      SectionContext c,
      FromEvent final Object model,
      Prop....) {

    return ViewRenderInfo.create()
        .viewCreator(VIEW_CREATOR)
        .viewBinder(new SimpleViewBinder() {
              Override
              public void bind(View view) {
                ((MyCustomView)view).bind(model)
              }
            })
        .build();
  }

```

In RecyclerView we keep the mapping between viewcreators and viewtypes to determine which viewcreator to use when adapter asks for it.

Reviewed By: pasqualeanatriello

Differential Revision: D5619305

fbshipit-source-id: b129d8e0380a35827697f8210aa414f4b509f8f4