Better error messages for text matchers
Summary:
Instead of just saying "Expecting: <com.facebook...> to have <>", we now output
something like this:

```
java.lang.AssertionError:
Expecting:
 <com.facebook.samples.litho.lithography.DecadeSeparator$DecadeSeparatorImpl@66c83fc8>
to have:
 <sub component <com.facebook.samples.litho.lithography.DecadeSeparator$DecadeSeparatorImpl@66c83fc8> with <expected = <is "2019">, actual text = <null>>>
```

The "actual text = <null>" is still bad, I know. But this is because we don't
know what we should match against in many cases. E.g. if we want to find *any
one* sub component that matches a given value, it's not clear what the "actual
value" is. A pretty-printed component tree we matched against might be the best
option - or a list of all compared values.

Anyway, this does already provide way more context than before.

Reviewed By: IanChilds

Differential Revision: D5641629

fbshipit-source-id: 177c16bee0f18fc4ca5fbfeabb8d391c0538f2da