Fix for not processing visibilityOutputs with transient state
Summary: When we have transient state on a LithoView, we want to mount everything within the view, but we don't want to process visibility events (since they might lie to us). The previous solution didn't work for children of the LithoView, since they still had no idea that we had transient state on the parent. Instead, let's pass a boolean around (uglier, but more correct).

Reviewed By: marco-cova

Differential Revision: D5649568

fbshipit-source-id: 567a0138e605c61a164a4153d1b5a5722bda7ca8