Parallelize maven downloads
Summary:
As JonCoens pointed out before, there is no good reason for the download process
not to run in parallel. My previous worry was that Maven would not properly lock
its repository in the file system, but for our case here this is easily avoided
by setting up a separare local repository for each instance in a temporary
folder.

The output is still interleaved and quite noisy, but you should only have to
read it if something went wrong and the last message will clearly indicate
wether or not it was successful (in addition to the exit code).

Reviewed By: JonCoens

Differential Revision: D5658713

fbshipit-source-id: 8f16b757827030f16278d0c4cc46dc790cdce9ac