Visible bounds takes full Litho hierarchy into account
Summary:
See comments on D5623667. The issue there is that we only look at the parent to determine how much of the LithoView is visible. However, it might be that the bounds of the parent are also not fully visible, in which case we need to find out how much of the parent is visible. We do this up to the top LithoView in the hierarchy (we can't go beyond this as we don't have control and we might end up never mounting ViewPager content, see D3893211).

Credit to stepanov for finding the bug here and suggesting the solution.

Btw I'm open to ideas about how to make this tidier - couldn't really think of a better way to do it.

Reviewed By: marco-cova

Differential Revision: D5649700

fbshipit-source-id: 16bc0bf29dc89a5d088eb8eaa2c9e8d917ddce70