Limit inheritance for annotation processing (1/2)
Summary:
== Why ==

I am going to add a new generator for `TestSpecs` to the existing
ComponentsProcessor. I personally follow the 0-1-2-n rule which allows
specializing for cases of 0, 1 and 2, but requires generalization for more. The
current implementation of AbstractComponentsProcessor is specialized to
LayoutSpecs and MountSpecs (i = 2), but adding my TestSpec generator would break
the rule.

== How ==

I want to follow the open-close principle and remove the knowledge of which
annotations are supported entirely from AbstractComponentsProcessor and push it
down into the concrete implementations. Secondly, I convert some of the
inheritance that was previously used for configuration to composition.

The concrete implementations are now fully self-contained and specify exactly
the annotations they support.

In order to support this, I moved the DependencyInjectionHelper out of
ComponentsProcessorUtils, (having less stuff in a `*Utils` class should be
fairly non-controverial) and wrapped a factory around it so we can use it in the
constructor. Similarly, the `*SpecModelFactory` classes now implement a common
interface.

Reviewed By: IanChilds

Differential Revision: D5687552

fbshipit-source-id: b64a16afa11d596bb433592ceb7cd2c370be74ba