Reset focused range when recycling VisibilityItems
Summary:
`VisibilityItem.mFocusedFlag` was not being reset in `release()`, which was leading to bugs when the item was recycled (we thought that the focused event had already been called, when in fact it had not). Instead, let's just use `mFlags` which is already being correctly reset.

I also wondered why we would recycle items with them being in a focused state. This wasn't a bug as such, since the items were no longer going to be used, but now the code in `MountState` also leaves them in the correct focused state (i.e. unfocused) before releasing them. Previously, it would only do this if the item had an unfocused event.

Reviewed By: muraziz

Differential Revision: D5726333

fbshipit-source-id: 3035f47c6debbe51fb3765c93713de45bcf035da