Autogenerate unique key for components
Summary:
Try to set a unique global key on sibling components of the same type if no manual key is set.
Each parent component holds onto a map from the type of children it has to a counter that was appended to the child's global key. This should help us set unique keys on components.
The fallback of this approach is that if the order of the children changes, the key is not stable so info that relies on the component's key, such as state, will be lost. In this situation the user would have to set a manual key on the components that are expected to be moved around.

Reviewed By: pasqualeanatriello

Differential Revision: D5648922

fbshipit-source-id: 0cecf73b8f86bd7727438617ebbf5ba5cd3524f6