Don't process visibility outputs with transient state
Summary:
When the LithoView has transient state we mount everything, even if incremental mount is enabled, since we may not get the relevant hooks to mount content later on (while the view is animating). We also don't process visibility outputs while we have transient state (if we are mounting everything, then we could fire focused events that are incorrect). Instead, we just process them when the animation is finished, as a sort of "best guess".

This case was omitted when the processVisibilityOutputs code was added.

Reviewed By: pasqualeanatriello

Differential Revision: D5735770

fbshipit-source-id: ae5a8304b73ee2f5c42961b4d68192cd229e679b