Fix broken ComponentTreeIncrementalMountTest
Summary:
In OSS CI, we run tests in both debug and release mode. This test, however, only
passes in debug mode. I've added better catching of the expected exception and
skip the test if we're not in an internal build.

CI broke with D5735980 / c62fcae6b9080baa56fea2d543659166ee9fce67.

Reviewed By: IanChilds

Differential Revision: D5746270

fbshipit-source-id: ccc554e29ae89153bb895f44ed76438e7eda803e