Don't change ViewType mapping in RecyclerBinder when updated item is same type as existing one.
Summary:
To keep track of ViewType to ViewCreator mappings during `updateItemAt()` operation in `RecyclerBinder` we were removing `ViewCreator` of olditem and adding `ViewCreator` of new item to the map if the usage count was 1. While it works well to keep correct mappings of `ViewType` to `ViewCreator` and their usages there is an issue with `RecyclerView` animations.
If during a layout an item's ViewType has changed, `RecyclerView` will trigger add/remove animation even when the new item being updated is same as old one, which results in unnecessary fade-in/fade-out animation.
In this diff I made sure not to update ViewCreator usage and ViewType mappings if both old and new item's ViewCreator are same in updateItemAt() call.

Reviewed By: mihaelao

Differential Revision: D5737648

fbshipit-source-id: aa55641819c65326711d2596e40418db643e482e