Fix javadoc for processor
Summary:
There was actually a reason why we had a dependency from processor to core. And
it's a seemingly silly one: javadoc. Javadoc uses the classpath during
generation and requires all at-linked-to references to be available. Sadly, this
means that we can no longer cross-reference from the processor javadoc to core.

Another change came aparently with java8: You can't document an exception with
`throws` if it isn't explicitly marked as checked exception from that method.

Reviewed By: IanChilds

Differential Revision: D5784065

fbshipit-source-id: a53178e50613f9ea77f89e608e621da3923f242a