Add new models for Diff params
Summary:
These models will be used as described - for props or state that are of type `Diff`. We need a special model for these as they are the same as props and state except that the type is wrapped. The model knows this. This means that e.g. we can extract props from the model just by getting the underlying prop.

I removed ShouldUpdateValidation as part of this change. I could have fixed it up to match the new models but it would be removed further up the stack anyway and didn't seem worth the effort. Since it's just validation, it doesn't actually make this diff break without the diffs further up the stack, so I don't see an issue.

Reviewed By: passy

Differential Revision: D5706349

fbshipit-source-id: 6eafca7157a9b96239443f2ee3cd659171b54ba1