Tooltip Api
Summary:
Api for displaying a tooltip on a component anchor without using viewtags.
This provides a way to show a tooltip anchored to the predefined positions in TooltipPosition, with the possibility of setting an offset from those positions.

The reason that the ComponentTree requires a Tooltip implementation instead of a PopupWindow is that it will also be used with custom tooltip implementations.

Example usage: P58133579

Reviewed By: IanChilds

Differential Revision: D5766486

fbshipit-source-id: b518e0890b619f72e12cba0671e2901d4d62ae5e