Initial boilerplate for TestSpec generation
Summary:
This is some of the initial boilerplate for generating `TestSpec`s. I am going
for an incremental approach for landing this which will include a bunch of
unfinished and frankly useless code and will refactor along the way.

Please still feel free to point out rough corners in the reviews. I want to make
sure to at least leave TODOs in the relevant places.

For a more general outline, see https://fb.quip.com/WJgwAti5brD6

Reviewed By: IanChilds

Differential Revision: D5754875

fbshipit-source-id: 0cdc633f06d24f6b9fdd6997a38cd642ed2c3f81