Use node component for visibility outputs
Summary: Previously we were using the Component where the handler was defined. The component is used for two things - global key and output id. It feels that it was a bug for calculating output id as you might have set the same handler on a different component. Global key should still work fine as keys are passed down through the hierarchy so if it had changed before then it will have changed now too (and could have additionally changed now when it didn't before).

Reviewed By: mihaelao

Differential Revision: D5793428

fbshipit-source-id: 4f2c0980a64029b73fb20719a198b7fb30d89f0a