Introduce Mock Models for testing
Summary:
The specmodels infra gives us the wonderful benefit of having a lot of POJOs to
work with. However, we don't fully utilize this in the tests as we still mock
various simple getters. This obscures logic and reduce coverage of the
underlying logic.

For my TestSpec tests, I would like to avoid Mockito-style mocking where
possible and work directly with the interfaces instead. I have created a couple
of Model Models that come without any validation and builders that map 1-to-1 to
the underlying fields which makes setting up objects for tests super easy.

== Follow-Up ==

There are two items I will address in follow-up diffs:

- BUCK module size
- Duplication with TestMethodParams

The latter in particular is something I noticed too late and addressing it in the same diff introduces too much noise. Temporary duplication shouldn't be a big deal.

Reviewed By: IanChilds

Differential Revision: D5785311

fbshipit-source-id: f17576ea73e534205233c4a18ac3b47643ce63ca