Fix javadoc
Summary:
Whoever designed javadoc must have done it with the intent of creating an excuse
for not writing documentation.

This. Is. The. Worst.

It compiles now, and should render roughly in a way you would it expect to. The
weird whitespace and seemingly randomly placed `{code}` blocks are just there
to silence errors or actually create the rendered output you would expect. Sigh.

Reviewed By: IanChilds

Differential Revision: D5813704

fbshipit-source-id: b5ca7c247f892377ac2a98e7ab280c74e2b64848