Viewport change from view added without scrolling will trigger a onViewportChanged event
Summary:
The first data load does not trigger a tail fetch even those if the items do not fill up the viewport.

This diff will address this bug because when views are added on the visible range of the viewport, `onViewportChanged` will be triggered, and the logic for tail fetch will be subsquently processed.

Reviewed By: pasqualeanatriello

Differential Revision: D5841645

fbshipit-source-id: 715fc49d9013587e1cdb9afa63886e942a8fbb38