Use SparseArray in LayoutAttributes object
Summary:
It seems that the current implementation of the LayoutAttributes object is too heavy, which causes problems as it is used so often.

Let's try using a SparseArray to store the attributes. This should use considerably less memory but there is a concern over CPU usage as we're now doing a bit more work than before - and we're unboxing, boxing etc etc.

Reviewed By: pasqualeanatriello

Differential Revision: D5833813

fbshipit-source-id: 8d936e97841769bd546bc597fbbba1578901bd01