Fix the loading state issue with search section components
Summary: Previously, the loading spinner only shows up the first time it loads, since isEmpty is true initially. After data is loaded, the var is set to false. But its never set to true when loaded again. Thus the spinner never shows up again. The same with error and null states. Fixing that

Differential Revision: D5826743

fbshipit-source-id: 319f019729d356e96b34b0ac077915adca416506