Revert the revert of class annotation retention
Summary:
Reverts the revert of a revert of a revert of ... which was originally D5527203.

The diff was reverted because it caused some visual regressions. While I
couldn't reproduce the issue reported after reapplying the patch, it became
obvious when comparing the `buck-out/annotations/` output that this must have
been caused by missing `ContextScoped` annotations, causing new allocations of
objects that should have been kept around: P58241334

I've now added some ugly conditional logic to
FBDependencyInjectionHelperFactory which retains the original behavior while
ensuring that we do not emit repeated annotations, as this is unsupported on
Java 7 (and generally undesirable).

The new `diff -ur` output confirms that apart from the order, semantically there
is no difference for the scoping annotations: P58241492

NB: I'll ignore the Java style/BUCK warnings for now to make sure this will continue to revert cleanly if needed.

Reviewed By: IanChilds

Differential Revision: D5833667

fbshipit-source-id: 832a406344aab7de72ba9560fb8f212831987671