Add BUCK target to compile test spec sources
Summary:
This add yet another hack to our Component rules. When `test_spec_srcs` is
specified for a `litho_android_library`, this creates a new hidden
`<name>_testspec` target which compiles the provided sources with javac 8 and
the `-parameters` option enabled. This allows us to read the method parameter
names from the `.class` files generated.

This is not meant to be a permanent solution, but to unblock me while
investigating options to persist parameter information across compilation runs.

Reviewed By: emilsjolander

Differential Revision: D5853965

fbshipit-source-id: 409ce24cfc7960777cd1824e482d289f09ea3e7d