SectionLifecycle Temporary Method for Last Item Attached
Summary:
LocalSurface is unable to fetch more data when the first fetch does not fill up the entire viewport.

`OnViewportchanged()` is not triggered because there is no change in the total number of items and their visible positions.

This code changes in this diff is only temporary to unblock the product team.
I have set an optional prop to only allow the changes to affect LocalSurface product.

Differential Revision: D5860922

fbshipit-source-id: dba5d3df8e5a80b4863bc88d5d9d31946a5d8b0f