Fix Border Drawing when Bounds are not at (0,0)
Summary: If the bounds of the border drawable were specified to something other than (0,0) then the result was essentially garbage. Now we accomodate for that by first translating the canvas to wherever the position of the bounds are and then setting the draw bounds to (0,0) so the draw code works as expected.

Reviewed By: sjkirby

Differential Revision: D5867029

fbshipit-source-id: a9242c048bb3b21e6541f2d6df47af5210c962e4