Get correct rect for incremental mount and workaround for viewpagers
Summary:
Using `mLithoView.getLocalVisibleRect` when doing incremental mount is the correct thing, since it will determine exactly what is on the screen and only mount that, which is what we want. We used to use this when we first wrote incremental mount but unfortunately we came across a problem with ViewPagers, which is that they do not give any hooks to their children when they change page, so if some content was laid out on a non-visible page then we'd never mount it, which is visually pretty bad. We fixed that by just comparing to the immediate parent, but that is leading to bugs with visible events (since we mount when things aren't actually visible).

The approach with this fix is to mitigate the problems that we see with ViewPager. We do this by traversing the hierarchy to find any ViewPagers that are ancestors of the LithoView, and if we find any then we attach a listener to itso that we will do incremental mount when the ViewPager is scrolled.

N.B. We had discussed manually setting the listener on the ViewPager. However, it's very unclear to me how this listener could work efficiently. It could traverse every child and perform incremental mount if the child is a litho view, but that seems prohibitively expensive. Also, fixing this in the framework is preferable I would have thought?

Question:
- I'm adding the ViewPagerListener(s) once when we first incrementally mount, and then removing them when we clear the LithoView. Is that lifecycle correct, or is that going to miss some cases?

Reviewed By: marco-cova

Differential Revision: D5871392

fbshipit-source-id: 7c3e010463ba763b5f26e710f79b538fc146e905