Fix collecting animations at mount time
Summary:
Mount-time animations are transitions we collect when we go to mount a new LayoutState. Collecting these transitions are deferred because the onCreateTransition callback depends on Diff's of the previously mounted Props/State for that Component.

This was broken because we weren't collecting any animations at mount time unless there were also non-mount-time animations available. This diff collects those animations no matter what and merges them in with non-mount-time animations

Reviewed By: marco-cova

Differential Revision: D5843434

fbshipit-source-id: 73c86e4a0c926d9c3bda17614406023576ce449a