Only reset alpha if NodeInfo had alpha set
Summary:
This fixes the attached bug, but is going to make it so we start resetting alpha when we shouldn't at the end of animations.

The correct long term fix will be to be able to hold onto LayoutOutputs throughout the animation and reset to the end LayoutOutput when the animation ends, but that requires larger changes than I have time to do right now (I created a todo though).

Reviewed By: edelron

Differential Revision: D5964535

fbshipit-source-id: 580585f77c85c8df1c390b25dc63658c589f253f