Only preallocate one MountableContent at a time
Summary: MountContent preallocation will cause the entire pool to be filled immediately. With preallocation running on the UI thread (during idle, but still on the UI thread), we can easily lock up the UI thread for several seconds just preallocating Footers and Videos. We need to be careful about this preallocation. This only preallocates one item per job. This is definitely not the right end state, but it's a step in the right direction given what we have today.

Reviewed By: WaseemTheDream

Differential Revision: D5978691

fbshipit-source-id: c3831a1bbd4181bc1d5057ab7d51ba83712544fc