Allow duplicate manual keys to be set on Components
Summary:
This relaxes the condition that all manual keys set on sibling components should be unique. We will allow duplicates and apply the same unique key autogeneration on top of the manual key.
This could result in issues when using APIs that rely on knowing the local key of a Component, such as the tooltip API, but we don't expect this to result in major issues and we will give a warning message that this results in unexpected behaviour.

We won't need D5893565 anymore.

Reviewed By: IanChilds

Differential Revision: D5932197

fbshipit-source-id: 69ec277a265bab14d8180ad711293ed496b81637