Add height animation support for LithoView in a RecyclerView to the transitions API
Summary:
Adds support for height animations as the title says.

Skipping width for now just to keep this simple.

This uses the existing TransitionManager infra to drive the animation. This means that you can interleave a bounds animation using `Transition.sequence` or `Transition.stagger` and that the animation is interruptible.

In the API, you can cause a LithoView bounds animation by animating the height of the root component mounted in your LithoView.

The way it works is that when LithoView detects a height change, it checks if the LayoutState defines a height animation on its root component. If it does, instead of committing the new height, it keeps the old height and puts itself in a state where it expects its height to be animated. When we mount the LayoutState and run transitions, we the animate the LithoView's height.

The reason we can't just animate the LithoView's height is because transition animations don't start until mount time, which is after measure: if we use the updated height immediately, we'll cause the view's siblings to move and then move back when we start the animation, causing a flicker.

Reviewed By: marco-cova

Differential Revision: D5923504

fbshipit-source-id: 945ae72dfb89bc4dfc1a1911147702cf049a47a5