Add height animations for components within a LithoView
Summary:
Depends on D5923504

This is similar to D5923504 but instead of just animating the bounds of the LithoView, we are animating components within that LithoView. The main difference here in implementation beyond the other animation types is that when we set a transitionKey on a component, we automatically wrap it in a View (ComponentHost). This means when we animate the bounds of the ComponentHost, we don't automatically animate the bounds of the drawable that we originally wanted to animate. That's why when animating bounds, we now automatically animate the "linkedDrawableForAnimation" which is just the drawable corresponding to the component that originally had a transitionKey set on it. The assumption is that there will only ever be one of these Drawables in a ComponentHost which is definitely the case for now because we always wrap in a unique ComponentHost.

There are issues with this API, namely that animating the height on a component doesn't automatically animate its siblings. Since we aren't actually running layout, this isn't possible to do right now. Longer term, I want us to move to a more "magic-move" transition API where you don't have to explicitly animate specific components when doing layout changes. For now, I'd like to get this committed to unblock feed.

Reviewed By: marco-cova

Differential Revision: D5934181

fbshipit-source-id: cef1f4f9084361ebf129c7942f2989c0f8046320