Generics on the Builder.create
Summary:
The BuilderPools and the Builder class in each component is now always static instead of depending on whether the Component class has generics or not.
In case of a generic that is also expected as a Prop, you might either need to specify it as `MyComponent.<MyType>create(c)` or in case of a Prop accepting a `List<MyType` you can avoid the generics on the .create() and use just a List type as a prop getting an unchecked cast warning.

Reviewed By: pasqualeanatriello

Differential Revision: D5973932

fbshipit-source-id: a6996661da5d152d0804988c8d848278e26bd144