Add model element type
Summary: This is to distinguish between sources that specs are generated from. So far, we'll support standard java classes and kotlin-style singletons. I'm not entirely sure if this is actually the best (or even correct) way to detect Kotlin `object`s, but it works and we can change it if we find a better way.

Reviewed By: pasqualeanatriello

Differential Revision: D5996268

fbshipit-source-id: c2fa16e4646eb88342871823e16240f1728eb7e1