fix stale litho selendroid elements issue
Summary:
An issue, where element returned to selendroid client, is released by litho (UI refresh) then the test item is reused for another element.
We see it with clicking the wrong element during a test, but it can also manifest in other ways like not finding the element that is on the screen, getting the wrong text, capturing the wrong screenshot, etc.

I don't see a way to solve it in litho because on click it doesn't know how the element was found to find it again.

So my solution is
1. Detect an element is stale by using acquire-key.
2. Check in a bunch of places if the element is stale and throw an exception if so.
3. On e2e test side we execute `findElement` - `doAction` in an atomic method that handles the stale exception and retries.

This requires that all `WebElement` usages be wrapped in `StaleElementHandler`
- I added `LITHO_STALE_DETECTION_ENABLED` flag so the stale detection can be active only for test that can handle it.
* Relatively easy to do for fblite as we almost never call `findElement` directly.

Any better ideas?

Reviewed By: marcelogomez, passy

Differential Revision: D6020809

fbshipit-source-id: ffc4504c1c4c08ef64d625430f440aa9c8445d27