Clean up preallocateMountContent
Summary:
Depends on D6081368
There's an unnecessary call to `removeCallbacks` in the chain of events of preallocateMountContent. Due to the implementation, each ComponentTree gets 1 runnable. Those ComponentTrees are reused, so the Runnables are reused. And posting that Runnable into the IdleExecutorHandler will mean the `mRunnablesHashMap` inside the IdleExecutorHandler will keep overwriting the mapping from key to value. The underlying executor may behave differently, but canceling the runnable in the queue will not have the same effect as actually canceling other runnables.

tl;dr this is an unnecessary couple lines of code

`preallocateMountContent` probably needs a rewrite, but for now let's keep it as clean as possible

Reviewed By: mihaelao

Differential Revision: D6081845

fbshipit-source-id: 73199985d1301e13717d16e9e982140ed954d285