Add parent when creating LithoViews in tests
Summary: Prior to this change to Android: https://sourceforge.net/p/android-x86/frameworks_base/ci/002d43d29b84e8a4e55c995a350783e80132d61a/, if a ViewGroup had no parent then when calling getChildVisibleRect we would update the offset for the child but not the visible rect. This isn't a bug as such as we will always have a ViewRootImpl with a single child when running real code, but in tests we just had a LithoView with no parent, leading to incorrect values being returned from getGlobalVisibleRect. This solves that problem.

Reviewed By: passy

Differential Revision: D6134339

fbshipit-source-id: f2d4258da978cd50711d844953ba0103ab3ef4de