Update OSS tutorial to use section components.
Summary:
THis diff updates the OSS tutorial to use section components! It's a follow up to D5944003 and updates the doc to reflect code updates.

I'm doing this here the make the change easier to review since it diffs with the old tutorial. I also haven't completed it because it's missing screenshots.  I think I got all the links right though.

Reviewed By: mihaelao

Differential Revision: D5987180

fbshipit-source-id: e5a540f2b099af02b61f85de84c5be14bb2e69c1