Collect transitions from delegating components
Summary: Previously we'd only collect transitions from the component returned from `InternalNode.getRootComponent()`. With this diff we also collect transitions from delegate components. Note, while we don't have or expect people to write components that define transitions on both the root component and on delegating components, that should work with this diff now.

Reviewed By: marco-cova

Differential Revision: D6124799

fbshipit-source-id: 39f0bec693ab336f7a62765787575dde94fcc4e0