Make StateContainer implementations @VisibleForTesting
Summary:
Sometimes when testing a component you want to assert on that component's state.  This is impossible right now because the StateContainerImpls are private.   This diff updates the annotation processor to make StateContainerImpls package-private instead and marks them as VisibleForTesting.

NOTE: This makes the annotation processor bring in android support annotations because that's where VisibleForTesting lives. Idk if that's kosher or not.

Reviewed By: boxme

Differential Revision: D6134973

fbshipit-source-id: 696d5bacc675a0a3a78e2a97d98b4b68841f4131