Set animated property to final value when animation finishes
Summary:
Previously we would reset the property at the end of the animation. This works for things like translate, but not when animating to a statically set alpha or scale value (since we can't just reset it to 1).

This diff makes it so we keep around the LayoutOutput until the end of the animation and use it to set a final value at the end of the animation.

Reviewed By: marco-cova

Differential Revision: D6134556

fbshipit-source-id: 6af91f51b911ea6ee63c43a49e640e7adfb74f06