Do not remove ViewCreator-ViewType mappings when no more usage left
Summary:
When providing `ViewCreator` via `ViewRenderInfo` we were keeping mapping of unique `ViewCreator` to RecyclerView Adapter's viewtype (we just increment viewType for each new viewcreator). Also to save memory we were keeping usage count and removing mapping when no more item was left in adapter that corresponded to its ViewCreator. While that works well in most scenarios where `ViewRenderInfo` is used as temporary solution and just a fraction of items in list are created by `ViewRenderInfo`, it is not efficient with scenarios where all items are created by `ViewRenderInfo`. If you for example remove items of same type until usage goes to 0, we remove it and if we add such item back, its viewtype changes where we lose the ability to re-use previously inflated views.

This diff removes the usage count so that the viewType to ViewCreator stays same throughout the adapter lifecycle. This will make sure we still use views from RecyclerView Recycler's pool even when we removed all items and added new items with similar types. While it might add some memory footprint, the ViewCreator instances should be light and generally don't store any local variables, so it shouldn't affect performance much.

Reviewed By: mihaelao

Differential Revision: D6147433

fbshipit-source-id: aa86922333ffd4b6e9578a6fb54948464a7c5d79