Prevent State Loss in Transition Demo App
Summary:
State is lost on rotation even when the `LithoView` in `TransitionsDemoActivity` is not recreated. This is because we did not implement a `sameItem` check for the `DataDiffSection` in `StoryFooterComponent`. That's why it removes and then inserts a new row on a new device rotation, which causes state loss.

We implement the check now in `TransitionsDemoComponent`

Reviewed By: mihaelao

Differential Revision: D6182659

fbshipit-source-id: e4d6dc19f35ec163dda650d322c7e49f45562a8a