Fix @PropDefault for Kotlin
Summary:
Before this commit, `PropDefault` would not get picked up by our annotation
processor at all. This is because of the weird way in which KAPT generates its
Java stubs.

This diff gets us half-way there in allowing `PropDefaults` iff they are
annotated with `JvmField`. This restriction is necessary because otherwise, the
field is marked as private and requires access through an automatically
generated getter method. We could fix this in the future, but this is enough to
unblock us for now.

I will document the limitation in the GitHub issue.

Reviewed By: pasqualeanatriello

Differential Revision: D6187360

fbshipit-source-id: 75a5499f1fab5771b06a94ff504970b4f31a1229