Restructure testing module
Summary:
As part of the last diff, I had to move `LithoRepresentation` up one
level from where it should sit - under assertj. That was because of the testing
mega-module. This attempts to untangle things a bit and create more smaller
modules. This should generally help build speed. And this is a highly
depended-upon module.

Reviewed By: IanChilds

Differential Revision: D6198838

fbshipit-source-id: e058b63a0ac85fcf516cc71f6498058eb066dd56