Enable mount content preallocation per spec
Summary:
in multirow we use the part definition to decide whether a ComponentTree can preallocate mount specs or not. With sections we won't be able to do that anymore. We want to be able to set on the spec whether it can be preallocated or not.
This adds a param to MountSpec to enable/disable preallocation. We'll check whether this is enabled to preallocate spec objects instead of preallocating all mount specs for a component tree.

This diff doesn't change the behaviour of the current preallocation implementation. I will start using this behind a qe for sections.
Will later add a method for configuring this at runtime.

Reviewed By: IanChilds

Differential Revision: D6210147

fbshipit-source-id: 97f6aab6f9e7a396ee0ca73b9bf7199e8c4a9961