Add JUnit @Rule and Activity for Litho screenshot tests
Summary:
The general idea is that animation testing isn't scaling: they're being added to this TransitionsDemoActivity and then trying to find the right row and all this crap -- it's not good. I want to encourage people (and myself) to write more tests for animations and I think a first step is allowing them to write separate tests instead of glomming them all into the same class 😛

So what that diff does is create a new Rule and Activity. The Rule lets you define a root component for that Activity, and then you can get the root LithoView to run your test. The nice thing here is to create a new test, you don't have to create a new Activity or mess with manifests.

To use it, you just add a LithoComponentActivityRule to your test with the component you want rendered, and it'll start an activity hosting that component for each test.

Reviewed By: passy

Differential Revision: D6212197

fbshipit-source-id: 8c9e473c24320f14488b6f8201c31f51ab8cfcd4