EventTrigger Outside of Component Hierarchy
Summary:
Components that declare `onTrigger` methods will be able to generate an `EventTrigger` object for trigger outside of the Component Hierarchy.

This is only necessary if user wants to trigger a component from outside of the hierarchy. For example, this allows us to trigger `RecyclerCollectionComponent` to `requestScrollToPosition()`, which in turn will replace the usage of `RecyclerEventsController`.

The APIs exposed to the users remain relatively unchanged, so there is no need to change existing usages in the products.

Reviewed By: mihaelao

Differential Revision: D5900370

fbshipit-source-id: 8324c791d492d84ba7a54ff01314515d15f5ac9f